Move metrics service
